% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-basic.R
\name{as_embed}
\alias{as_embed}
\alias{as_wordvec}
\title{Word vectors data class: \code{wordvec} and \code{embed}.}
\usage{
as_embed(x, normalize = FALSE)

as_wordvec(x, normalize = FALSE)
}
\arguments{
\item{x}{Object to be reshaped. See examples.}

\item{normalize}{Normalize all word vectors to unit length?
Defaults to \code{FALSE}. See \code{\link{normalize}}.}
}
\value{
A \code{wordvec} (data.table) or \code{embed} (matrix).
}
\description{
\code{PsychWordVec} uses two types of word vectors data:
\code{wordvec} (data.table, with two variables \code{word} and \code{vec})
and \code{embed} (matrix, with dimensions as columns and words as row names).
Note that matrix operation makes \code{embed} much faster than \code{wordvec}.
Users are suggested to reshape data to \code{embed} before using the other functions.
}
\section{Functions}{
\itemize{
\item \code{as_embed()}: From \code{wordvec} (data.table) to \code{embed} (matrix).

\item \code{as_wordvec()}: From \code{embed} (matrix) to \code{wordvec} (data.table).

}}
\section{Download}{

Download pre-trained word vectors data (\code{.RData}):
\url{https://psychbruce.github.io/WordVector_RData.pdf}
}

\examples{
dt = head(demodata, 10)

embed = as_embed(dt)
embed

wordvec = as_wordvec(embed)
wordvec

df = data.frame(token=LETTERS, D1=1:26/10000, D2=26:1/10000)
as_embed(df)
as_wordvec(df)

}
\seealso{
\code{\link{load_wordvec}} / \code{\link{load_embed}}

\code{\link{normalize}}

\code{\link{data_transform}}

\code{\link{data_wordvec_subset}}
}
