% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{data_transform}
\alias{data_transform}
\title{Transform plain text data of word vectors into a compressed ".RData" file.}
\usage{
data_transform(
  file.load,
  file.save = NULL,
  encoding = "UTF-8",
  sep = " ",
  header = FALSE,
  compress = "bzip2",
  compress.level = 9
)
}
\arguments{
\item{file.load}{File name of raw data (must be plain text).

Data must be in this format (values separated by \code{sep}):

cat 0.001 0.002 0.003 0.004 0.005 ... 0.300

dog 0.301 0.302 0.303 0.304 0.305 ... 0.600}

\item{file.save}{File name of to-be-saved R data (must be .rda or .RData).}

\item{encoding}{File encoding. Default is \code{"UTF-8"}.}

\item{sep}{Column separator. Default is \code{" "}.}

\item{header}{First row is header or meta-information? Default is \code{FALSE}.
If \code{TRUE}, then the first row will be removed.}

\item{compress}{Compression method for the saved file. Default is \code{"bzip2"}.

Options can be:
\itemize{
  \item \code{1} or \code{"gzip"}: modest file size (fastest)
  \item \code{2} or \code{"bzip2"}: small file size (fast)
  \item \code{3} or \code{"xz"}: minimized file size (slow)
}}

\item{compress.level}{Compression level from \code{0} (none) to \code{9}
(maximal compression for minimal file size). Default is \code{9}.}
}
\value{
A \code{data.table} with two variables: \code{word} and \code{vec}.
}
\description{
Transform plain text data of word vectors into a compressed ".RData" file.
}
\section{Download}{

Download pre-trained word vectors data (\code{.RData}):
\url{https://psychbruce.github.io/WordVector_RData.pdf}
}

