% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{data_wordvec_load}
\alias{data_wordvec_load}
\title{Load word vector data from an ".RData" file.}
\usage{
data_wordvec_load(file)
}
\arguments{
\item{file}{File name of the data (must be .rda or .RData transformed by
\code{\link{data_transform}}, with only two variables \code{word} and \code{vec}).}
}
\value{
A \code{data.table} with three variables: \code{word}, \code{vec}, \code{dim} (number of dimensions).

You may drop/delete the variable \code{dim} but should always keep \code{word} and \code{vec} in the data.
}
\description{
Load word vector data from an ".RData" file.
}
