% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{dict_reliability}
\alias{dict_reliability}
\title{Reliability analysis and PCA of a dictionary.}
\usage{
dict_reliability(
  data,
  words = NULL,
  pattern = NULL,
  sort = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A \code{data.table} (of new class \code{wordvec})
loaded by \code{\link{data_wordvec_load}}.}

\item{words}{[Option 1] Word strings (\code{NULL}; a single word; a vector of words).}

\item{pattern}{[Option 2] Pattern of regular expression (see \code{\link[stringr:str_subset]{str_subset}}).}

\item{sort}{Sort items by the first principal component loading (PC1)?
Defaults to \code{TRUE}.}

\item{plot}{Visualize cosine similarities with words ordered by PC1?
Defaults to \code{TRUE}.}

\item{...}{Other parameters passed to \code{\link{plot_similarity}}.}
}
\value{
A \code{list} object of new class \code{reliability}:
\describe{
  \item{\code{alpha}}{
    Cronbach's \eqn{\alpha}}
  \item{\code{eigen}}{
    Eigen values from PCA}
  \item{\code{pca}}{
    PCA (only 1 principal component)}
  \item{\code{pca.rotation}}{
    PCA with varimax rotation (if potential principal components > 1)}
  \item{\code{items}}{
    Item statistics}
  \item{\code{cos.sim}}{
    A matrix of cosine similarities of all word pairs}
}
}
\description{
Reliability analysis (Cronbach's \eqn{\alpha}) and
Principal Component Analysis (PCA) of a dictionary.
Note that Cronbach's \eqn{\alpha} may be misleading
when the number of items/words is large.
}
\section{Download}{

Download pre-trained word vectors data (\code{.RData}):
\url{https://psychbruce.github.io/WordVector_RData.pdf}
}

\examples{
d = data_wordvec_normalize(demodata)

dict = dict_expand(d, "king")
dict_reliability(d, dict)

dict.cn = dict_expand(d, "China", threshold=0.65)
dict_reliability(d, dict.cn)

dict_reliability(d, c(dict, dict.cn))
# low-loading items should be removed

}
\references{
Nicolas, G., Bai, X., & Fiske, S. T. (2021).
Comprehensive stereotype content dictionaries using a semi-automated method.
\emph{European Journal of Social Psychology, 51}(1), 178--196.
}
\seealso{
\code{\link{cosine_similarity}}

\code{\link{pair_similarity}}

\code{\link{plot_similarity}}

\code{\link{tab_similarity}}

\code{\link{most_similar}}

\code{\link{dict_expand}}
}
