% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{get_wordvec}
\alias{get_wordvec}
\title{Extract word vector of a single word.}
\usage{
get_wordvec(data, word)
}
\arguments{
\item{data}{Data table (\code{data.table}) with variables \code{word} and \code{vec}.

See \code{\link{data_transform}} for details about data format.}

\item{word}{Word string (a single word).}
}
\value{
A numeric vector of the word (or \code{NA} if the word is not in the data).
}
\description{
Extract word vector of a single word.
}
