% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{get_wordvecs}
\alias{get_wordvecs}
\title{Extract word vectors of multiple words.}
\usage{
get_wordvecs(data, words = NULL, pattern = NULL)
}
\arguments{
\item{data}{Data table (\code{data.table}) with variables \code{word} and \code{vec}.

See \code{\link{data_transform}} for details about data format.}

\item{words}{[Option 1] Word string (\code{NULL}; a single word; a vector of words).}

\item{pattern}{[Option 2] Pattern of regular expression (see \code{\link[stringr:str_detect]{str_detect}}).}
}
\value{
A \code{data.table} with words as columns and dimensions of word vectors as rows.
}
\description{
Extract word vectors of multiple words,
using either wordlist (a vector of words; using \code{words})
or regular expression (a pattern of words; using \code{pattern}).
If both (\code{words} and \code{pattern}) are specified, \code{words} wins.
}
