% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{more_similar_than}
\alias{more_similar_than}
\title{Find all words more similar to a word than a critical word is to the word.}
\usage{
more_similar_than(data, word, than)
}
\arguments{
\item{data}{Data table (\code{data.table}) with variables \code{word} and \code{vec}.

See \code{\link{data_transform}} for details about data format.}

\item{word}{Word string (a single word).}

\item{than}{Critical word (a single word).}
}
\value{
A \code{data.table} with all words with cosine similarities higher than
that of the critical word.
}
\description{
Find all words more similar to a word than a critical word is to the word.
}
