% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{most_similar}
\alias{most_similar}
\title{Find the Top-N most similar words.}
\usage{
most_similar(data, word, topn = 10, above = NULL)
}
\arguments{
\item{data}{Data table (\code{data.table}) with variables \code{word} and \code{vec}.

See \code{\link{data_transform}} for details about data format.}

\item{word}{Word string (a single word).}

\item{topn}{Top-N most similar words. Default is \code{10}.}

\item{above}{A threshold of cosine similarity to get all words with
similarity above this value. Default is \code{NULL}.
If both \code{topn} and \code{above} are specified, \code{above} wins.}
}
\value{
A \code{data.table} with the most similar words and their cosine similarities.
The row number of each word in the raw data is also returned,
which may help determine the relative word frequency in some cases.
}
\description{
Find the Top-N most similar words.
}
