% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{pair_similarity}
\alias{pair_similarity}
\title{Compute cosine similarity/distance for a pair of words.}
\usage{
pair_similarity(data, word1, word2, distance = FALSE)
}
\arguments{
\item{data}{Data table (\code{data.table}) with variables \code{word} and \code{vec}.

See \code{\link{data_transform}} for details about data format.}

\item{word1, word2}{Word string (a single word).}

\item{distance}{Compute cosine distance instead?
Default is \code{FALSE} (cosine similarity).}
}
\value{
A value of cosine similarity/distance.
}
\description{
Compute cosine similarity/distance for a pair of words.
}
\examples{
pair_similarity(demodata, "China", "Chinese")

}
