% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{plot_wordvecs}
\alias{plot_wordvecs}
\title{Visualize word vector(s).}
\usage{
plot_wordvecs(data, words, dims = NULL)
}
\arguments{
\item{data}{Data table (\code{data.table}) with variables \code{word} and \code{vec}.

See \code{\link{data_transform}} for details about data format.}

\item{words}{[Option 1] Word string (\code{NULL}; a single word; a vector of words).}
}
\description{
Visualize word vector(s).
}
