% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWV1_static.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cc}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bruceR}{\code{\link[bruceR]{cc}}}
}}

