% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWV2_dynamic.R
\name{text_env_setup}
\alias{text_env_setup}
\title{Set up a Python environment for contextualized (dynamic) word embeddings,
necessary for all \code{text_*} functions.}
\usage{
text_env_setup(install = TRUE, initialize = TRUE, save = TRUE, ...)
}
\arguments{
\item{install}{Run \code{\link[text:textrpp_install]{text::textrpp_install()}}.
Defaults to \code{TRUE}.}

\item{initialize}{Run \code{\link[text:textrpp_initialize]{text::textrpp_initialize()}}.
Defaults to \code{TRUE}.}

\item{save}{Save settings to .Rprofile for future use,
so that you do not have to run this again after restarting R.
Defaults to \code{TRUE}.}

\item{...}{Other parameters passed to
\code{\link[text:textrpp_install]{text::textrpp_install()}}.}
}
\description{
Install required Python packages in a conda environment (with defaults)
and initialize the installed conda environment.
Run once before your first use of any \code{text_*} function.
No need to run this in the next time.
Users may first need to manually install
\href{https://www.anaconda.com/}{Anaconda} or
\href{https://docs.conda.io/en/main/miniconda.html}{Miniconda}.
}
\details{
The R package \code{text} (\url{https://www.r-text.org/}) enables users access to
\href{https://huggingface.co/models}{HuggingFace Transformers models} in R,
through the R package \code{reticulate} as an interface to Python
and the Python packages \code{torch} and \code{transformers}.
This function calls two functions in the \code{text} package.

Users who have problems in running this function should directly
run the functions in \code{text}:
\itemize{
  \item{\code{\link[text:textrpp_install]{text::textrpp_install()}}}
  \item{\code{\link[text:textrpp_install]{text::textrpp_install_virtualenv()}}}
  \item{\code{\link[text:textrpp_uninstall]{text::textrpp_uninstall()}}}
  \item{\code{\link[text:textrpp_initialize]{text::textrpp_initialize()}}}
}
}
\examples{
\dontrun{
text_env_setup()
}

}
\seealso{
\code{\link{text_model_download}}

\code{\link{text_model_remove}}
}
