% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWV2_dynamic.R
\name{text_model_remove}
\alias{text_model_remove}
\title{Remove downloaded models from .cache folder.}
\usage{
text_model_remove(model = NULL)
}
\arguments{
\item{model}{Model name. See \code{\link{text_model_download}}.
Defaults to automatically find all downloaded models in the .cache folder.}
}
\description{
Remove downloaded models from .cache folder.
}
\examples{
\dontrun{
text_model_remove()
}

}
\seealso{
\code{\link{text_env_setup}}

\code{\link{text_model_download}}
}
