% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{utf8_split_default}
\alias{utf8_split_default}
\title{Default UTF-8 separators used to split words and sentences.}
\usage{
utf8_split_default()
}
\value{
A character vector of length 2:
(1) the first element indicates how to split words and
(2) the second element indicates how to split sentences.
}
\description{
Used only in \code{\link{train_wordvec}}.
}
\examples{
utf8_split_default()
}
