% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{data_wordvec_load}
\alias{data_wordvec_load}
\title{Load word vectors data from an ".RData" file.}
\usage{
data_wordvec_load(file.load, normalize = FALSE)
}
\arguments{
\item{file.load}{File name (must be .RData transformed by
\code{\link{data_transform}}, with only two variables \code{word} and \code{vec}).}

\item{normalize}{Normalize all word vectors to unit length?
Defaults to \code{FALSE}. See \code{\link{data_wordvec_normalize}}.}
}
\value{
A \code{data.table} with two variables:
\describe{
  \item{\code{word}}{words}
  \item{\code{vec}}{\strong{raw} \emph{or} \strong{normalized} word vectors}
}
}
\description{
Load word vectors data from an ".RData" file.
}
\section{Download}{

Download pre-trained word vectors data (\code{.RData}):
\url{https://psychbruce.github.io/WordVector_RData.pdf}
}

\seealso{
\code{\link{data_transform}}

\code{\link{data_wordvec_normalize}}

\code{\link{data_wordvec_reshape}}

\code{\link{data_wordvec_subset}}
}
