% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{data_wordvec_reshape}
\alias{data_wordvec_reshape}
\title{Reshape word vectors data.}
\usage{
data_wordvec_reshape(data, to = c("plain", "dense"), normalize = FALSE)
}
\arguments{
\item{data}{Data to be reshaped. See examples.}

\item{to}{Two options:
\itemize{
  \item{\code{"plain"} (default) reshapes the data
  from \code{data.table} (with only two variables \code{word} and \code{vec},
  loaded by \code{\link{data_wordvec_load}})
  to \code{matrix} (with dimensions as columns and words as row names).}
  \item{\code{"dense"} just does the reverse.}
}}

\item{normalize}{Normalize all word vectors to unit length?
Defaults to \code{FALSE}. See \code{\link{data_wordvec_normalize}}.}
}
\value{
A \code{data.table} (dense) or \code{matrix} (plain) of word vectors.
}
\description{
Reshape word vectors data (1) from dense to plain or (2) from plain to dense.
}
\section{Download}{

Download pre-trained word vectors data (\code{.RData}):
\url{https://psychbruce.github.io/WordVector_RData.pdf}
}

\examples{
d.plain = data_wordvec_reshape(demodata, to="plain")
d.plain

d.dense = data_wordvec_reshape(d.plain, to="dense")
d.dense  # identical to `demodata`

}
\seealso{
\code{\link{data_transform}}

\code{\link{data_wordvec_load}}

\code{\link{data_wordvec_normalize}}

\code{\link{data_wordvec_subset}}
}
