% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{data_wordvec_subset}
\alias{data_wordvec_subset}
\title{Extract a subset of word vectors data.}
\usage{
data_wordvec_subset(
  data = NULL,
  words = NULL,
  pattern = NULL,
  file.load.rda = NULL,
  file.save.rda = NULL,
  compress = "bzip2",
  compress.level = 9
)
}
\arguments{
\item{data}{Data table (\code{data.table}) with variables \code{word} and \code{vec}.

See \code{\link{data_transform}} for details about data format.}

\item{words}{[Option 1] Word string (\code{NULL}; a single word; a vector of words).}

\item{pattern}{[Option 2] Pattern of regular expression (see \code{\link[stringr:str_subset]{str_subset}}).}

\item{file.load.rda}{File name (must be .rda or .RData transformed by
\code{\link{data_transform}}, with only two variables \code{word} and \code{vec}).}

\item{file.save.rda}{File name of to-be-saved R data (must be .rda or .RData).}

\item{compress}{Compression method for the saved file. Default is \code{"bzip2"}.

Options can be:
\itemize{
  \item \code{1} or \code{"gzip"}: modest file size (fastest)
  \item \code{2} or \code{"bzip2"}: small file size (fast)
  \item \code{3} or \code{"xz"}: minimized file size (slow)
}}

\item{compress.level}{Compression level from \code{0} (none) to \code{9}
(maximal compression for minimal file size). Default is \code{9}.}
}
\value{
A subset of word vectors data of valid (available) words.
}
\description{
Extract a subset of word vectors data.
You must specify either \code{data} (a \code{data.table} loaded by \code{\link{data_wordvec_load}})
or \code{file.load.rda} (the name of an .RData file transformed by \code{\link{data_transform}}).
}
\section{Download}{

Download pre-trained word vectors data (\code{.RData}):
\url{https://psychbruce.github.io/WordVector_RData.pdf}
}

\examples{
## specify `data` and `words`:
d = data_wordvec_subset(demodata, c("China", "Japan", "Korea"))
d

## specify `data` and `pattern`, and then save the subset:
data_wordvec_subset(demodata, pattern="Chin[ae]|Japan|Korea",
                    file.save.rda="subset.RData")

## load the subset:
d.subset = data_wordvec_load("subset.RData")
d.subset

## specify `file.load.rda` and save with `file.save.rda`:
data_wordvec_subset(file.load.rda="subset.RData",
                    words=c("China", "Chinese"),
                    file.save.rda="new.subset.RData")
d.new.subset = data_wordvec_load("new.subset.RData")
d.new.subset

unlink("subset.RData")  # delete file for code check
unlink("new.subset.RData")  # delete file for code check

}
\seealso{
\code{\link{data_transform}}

\code{\link{data_wordvec_load}}

\code{\link{data_wordvec_normalize}}
}
