% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWV1_static.R
\name{get_wordvec}
\alias{get_wordvec}
\title{Extract the word vector of a single word.}
\usage{
get_wordvec(data, word)
}
\arguments{
\item{data}{A \code{\link[PsychWordVec:as_wordvec]{wordvec}} (data.table) or
\code{\link[PsychWordVec:as_embed]{embed}} (matrix),
see \code{\link{data_wordvec_load}}.}

\item{word}{Word string (a single word).}
}
\value{
A numeric vector of the word
(or \code{NA} if the word does not appear in the data).
}
\description{
Extract the word vector of a single word.
}
\section{Download}{

Download pre-trained word vectors data (\code{.RData}):
\url{https://psychbruce.github.io/WordVector_RData.pdf}
}

\examples{
d = as_wordvec(demodata, normalize=TRUE)

v1 = get_wordvec(demodata, "China")  # raw vector
v2 = get_wordvec(d, "China")  # normalized vector
cor(v1, v2)
cos_sim(v1, v2)

}
\seealso{
\code{\link{get_wordvecs}}

\code{\link{plot_wordvec}}

\code{\link{plot_wordvec_tSNE}}
}
