% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWV1_basic.R
\name{get_wordvec}
\alias{get_wordvec}
\title{Extract word vector(s).}
\usage{
get_wordvec(
  data,
  words = NULL,
  pattern = NULL,
  plot = FALSE,
  plot.dims = NULL,
  plot.step = 0.05,
  plot.border = "white"
)
}
\arguments{
\item{data}{A \code{\link[PsychWordVec:as_wordvec]{wordvec}} (data.table) or
\code{\link[PsychWordVec:as_embed]{embed}} (matrix),
see \code{\link{data_wordvec_load}}.}

\item{words}{[Option 1] Character string(s).}

\item{pattern}{[Option 2] Regular expression (see \code{\link[stringr:str_subset]{str_subset}}).
If neither \code{words} nor \code{pattern} are specified (i.e., both are \code{NULL}),
then all words in the data will be extracted.}

\item{plot}{Generate a plot to illustrate the word vectors? Defaults to \code{FALSE}.}

\item{plot.dims}{Dimensions to be plotted (e.g., \code{1:100}).
Defaults to \code{NULL} (plot all dimensions).}

\item{plot.step}{Step for value breaks. Defaults to \code{0.05}.}

\item{plot.border}{Color of tile border. Defaults to \code{"white"}.
To remove the border color, set \code{plot.border=NA}.}
}
\value{
A \code{data.table} with words as columns and dimensions as rows.
}
\description{
Extract word vector(s), using either a list of words or a regular expression.
}
\section{Download}{

Download pre-trained word vectors data (\code{.RData}):
\url{https://psychbruce.github.io/WordVector_RData.pdf}
}

\examples{
d = normalize(demodata)

get_wordvec(d, c("China", "Japan", "Korea"))
get_wordvec(d, cc(" China, Japan; Korea "))

## specify `pattern`:
get_wordvec(d, pattern="Chin[ae]|Japan|Korea")

## plot word vectors:
get_wordvec(d, cc("China, Japan, Korea,
                   Mac, Linux, Windows"),
            plot=TRUE, plot.dims=1:100)

\donttest{## a more complex example:

words = cc("
China
Chinese
Japan
Japanese
good
bad
great
terrible
morning
evening
king
queen
man
woman
he
she
cat
dog
")

dt = get_wordvec(
  d, words,
  plot=TRUE,
  plot.dims=1:100,
  plot.step=0.06)

# if you want to change something:
attr(dt, "ggplot") +
  scale_fill_viridis_b(n.breaks=10, show.limits=TRUE) +
  theme(legend.key.height=unit(0.1, "npc"))

# or to save the plot:
ggsave(attr(dt, "ggplot"),
       filename="wordvecs.png",
       width=8, height=5, dpi=500)
unlink("wordvecs.png")  # delete file for code check
}
}
\seealso{
\code{\link{data_wordvec_subset}}

\code{\link{plot_wordvec}}

\code{\link{plot_wordvec_tSNE}}
}
