% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychWordVec.R
\name{tab_similarity}
\alias{tab_similarity}
\title{Tabulate cosine similarity/distance of all word pairs.}
\usage{
tab_similarity(data, words, distance = FALSE)
}
\arguments{
\item{distance}{Compute cosine distance instead?
Default is \code{FALSE} (cosine similarity).}
}
\value{
A \code{data.table} of all combinations of the words,
with their wordpair and cosine similarity/distance.
}
\description{
Tabulate cosine similarity/distance of all word pairs.
}
\examples{
dts = tab_similarity(demodata, cc("king, queen, man, woman"))
dts

dts = tab_similarity(demodata, cc("Beijing, China, Tokyo, Japan"))
dts

}
