% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWV2_dynamic.R
\name{text_model_download}
\alias{text_model_download}
\title{Download pre-trained language models from HuggingFace.}
\usage{
text_model_download(model = NULL)
}
\arguments{
\item{model}{Character string(s) specifying the
pre-trained language model(s) to be downloaded.
For a full list of options, see
\href{https://huggingface.co/models}{HuggingFace}.
Defaults to download nothing and check currently downloaded models.

Example choices:
\itemize{
  \item{\code{"gpt2"}}
  \item{\code{"openai-gpt"}}
  \item{\code{"bert-base-uncased"}}
  \item{\code{"bert-large-uncased"}}
  \item{\code{"bert-base-cased"}}
  \item{\code{"bert-large-cased"}}
  \item{\code{"bert-base-chinese"}}
  \item{\code{"bert-base-multilingual-cased"}}
  \item{\code{"albert-base-v2"}}
  \item{\code{"albert-large-v2"}}
  \item{\code{"roberta-base"}}
  \item{\code{"roberta-large"}}
  \item{\code{"distilbert-base-uncased"}}
  \item{\code{"distilbert-base-cased"}}
  \item{\code{"distilbert-base-multilingual-cased"}}
  \item{\code{"xlm-roberta-base"}}
  \item{\code{"xlm-roberta-large"}}
  \item{\code{"xlnet-base-cased"}}
  \item{\code{"xlnet-large-cased"}}
  \item{\code{...} (see \url{https://huggingface.co/models})}
}}
}
\value{
Invisibly return the names of all downloaded models.
}
\description{
Download pre-trained language models (Transformers Models,
such as GPT, BERT, RoBERTa, DeBERTa, DistilBERT, XLM, etc.)
from \href{https://huggingface.co/models}{HuggingFace} to
your local ".cache" folder ("C:/Users/[UserName]/.cache/").
The models will never be removed unless you run
\code{\link{text_model_remove}}.
}
\examples{
\dontrun{
text_model_download()  # check downloaded models
text_model_download(c(
  "bert-base-uncased",
  "bert-base-cased",
  "bert-base-multilingual-cased"
))
}

}
\seealso{
\code{\link{text_env_setup}}

\code{\link{text_model_remove}}
}
