% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dynamic.R
\name{text_model_remove}
\alias{text_model_remove}
\title{Remove downloaded models from the local .cache folder.}
\usage{
text_model_remove(model = NULL)
}
\arguments{
\item{model}{Model name. See \code{\link{text_model_download}}.
Defaults to automatically find all downloaded models in the .cache folder.}
}
\description{
Remove downloaded models from the local .cache folder.
}
\examples{
\dontrun{
# text_init()  # initialize the environment

text_model_remove()
}

}
\seealso{
\code{\link{text_init}}

\code{\link{text_model_download}}

\code{\link{text_to_vec}}

\code{\link{text_unmask}}
}
